refactor allele_harvester to bmeg.enrichers
